Python/pyproject.toml
[tool.poetry]
name = "macro"
version = "0.1.0"
description = ""
authors = ["Anthony Diaz <anthony.gdmou@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyautogui = "^0.9.54"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
